Move common fixtures to conftest.py
There is a file specifically designed for this purpose.
We should use it.

It was used before but removed for some reason. Probably me.

This makes the linters happy without needing lots of ignore
or disable statements.
