+++ iati/core/tests/test_schemas.py
@@ -7,7 +7,6 @@ import iati.core.exceptions
 import iati.core.resources
 import iati.core.schemas
 import iati.core.tests.utilities
-from iati.core.tests.utilities import standard_version_optional
 
 
 class TestSchemas(object):
